build-multiarch: Disable cgo
The currently built Linux binary is not compatible with non-glibc Linux
distros such as Alpine, which do not contain glibc.

This is due to the use of cgo, which is enabled by default "for native
builds on systems where it is expected to work".

This change disables cgo by setting CGO_ENABLED=0 so that the Linux
build can run on different distros.

More info here: https://pkg.go.dev/cmd/cgo