publish-release: Support existing release
The publish-release workflow has been designed to create a release when
a tag is pushed, but the GitHub UI only allows users to create a tag by
creating a release. When this action creates a tag, it will trigger the
publish-release workflow to create a release, which already exists.

To address this, the workflow has been changed to check if a release
already exists. If it exists, the workflow uploads release artifacts to
the existing release. If not, the workflow creates the release as usual.