fix: --timeout delayed by --per-attempt-timeout (#13)
The total timeout (`--timeout`) does not abort during an attempt when
the `--per-attempt-timeout` is set. This can cause the real timeout to
greatly exceed the specified timeout.

This change fixes the behavior so that, when the total timeout elapses,
it will cancel any active attempt and exit with an error.

Fixes #13.